Replace unused Foreign Function module with Impl
