+++ unison-runtime/src/Unison/Runtime/Machine.hs
@@ -75,7 +75,7 @@ import Unison.Runtime.Array as PA
 import Unison.Runtime.Builtin hiding (unitValue)
 import Unison.Runtime.Exception hiding (die)
 import Unison.Runtime.Foreign
-import Unison.Runtime.Foreign.Impl (foreignCall)
+import Unison.Runtime.Foreign.Function (foreignCall)
 import Unison.Runtime.MCode
 import Unison.Runtime.Stack
 import Unison.Runtime.TypeTags qualified as TT
