+++ unison-runtime/unison-runtime.cabal
@@ -49,7 +49,6 @@ library
       Unison.Runtime.Foreign
       Unison.Runtime.Foreign.Function
       Unison.Runtime.Foreign.Function.Type
-      Unison.Runtime.Foreign.Impl
       Unison.Runtime.Interface
       Unison.Runtime.IOSource
       Unison.Runtime.Machine
